+ Profit Calculator - Fully working (I hope so, rng drops not tested yet)

+++ src/main/java/com/github/may2beez/farmhelperv2/handler/GameStateHandler.java
@@ -95,10 +95,10 @@ public class GameStateHandler {
     private BuffState godPotState = BuffState.UNKNOWN;
 
     @Getter
-    private long currentPurse = 0;
+    private double currentPurse = 0;
 
     @Getter
-    private long previousPurse = 0;
+    private double previousPurse = 0;
 
     @Getter
     private long bits = 0;
